Fix trap monsters disabling under mr4<
If under mr4 or lower a trap monster is disabled (via jinzo for example), the monster won't be able to return to its previous location, and will remain stuck in the monster zone, that because the zone locking effect is being reset on RESET_TURN_SET, but the adjust operation isn't refreshing the locations after the effects are reset (like it was doing before the mr5 update). Also, almost all the steps in trap_monster_adjust can be skipped if the duel isn't mr5 as there will always be a free location due to trap monster mechanics in previous master rules.