Pass extra parameter when retrieving the value of EFFECT_CHANGE_BATTLE_DAMAGE
This is needed mostly for Ghostrick Mansion, that card halves the battle damage only when a non ghostrick monster is inflicting it. The current script instead will halve the battle damage whenever a non ghostrick monster battles. The extra parameter, will be a boolean indicating if the handler of that effect is the card actually inflicting the damage. This will also require changing mansion's script from using a field effect to using grant effect, so that the handler will be properly set.